feat(resetPassword): separate views
- split newPassword view and resetPassword view

[#162949036]
